moving UI DTOs and UI mappers to view model
