+++ controller/src/main/java/com/khoshnaw/controller/movie/MovieController.kt
@@ -1,7 +1,6 @@
 package com.khoshnaw.controller.movie
 
 import com.khoshnaw.controller.base.Controller
-import com.khoshnaw.entity.Movie
 import com.khoshnaw.usecase.movie.loadMovieList.LoadMovieListInputPort
 import com.khoshnaw.usecase.movie.loadMovieList.LoadMovieListOutputPort
 import javax.inject.Inject
@@ -12,5 +11,5 @@ class MovieController @Inject constructor(
 
     suspend fun loadMoviesList() = inputPort.startUpdatingMovieList()
 
-    suspend fun showMovie(movie: Movie) = println("showing movie : $movie")
+    fun showMovie(id: String) = println("showing movie : $id")
 }
