--- ui/src/main/java/com/khoshnaw/ui/base/adapter/StateListItem.kt
@@ -1,3 +1,0 @@
-package com.khoshnaw.ui.base.adapter
-
-interface StateListItem
