+++ ui/src/main/java/com/khoshnaw/ui/mvi/MVIHolder.kt
@@ -2,7 +2,7 @@ package com.khoshnaw.ui.mvi
 
 import androidx.databinding.ViewDataBinding
 import com.khoshnaw.ui.base.adapter.BaseHolder
-import com.khoshnaw.ui.base.adapter.StateListItem
+import com.khoshnaw.viewmodel.dto.StateListItem
 
 abstract class MVIHolder<in I : StateListItem>(
     binding: ViewDataBinding
