+++ ui/src/main/java/com/khoshnaw/ui/standard/adapter/StandardHolder.kt
@@ -5,6 +5,7 @@ import androidx.lifecycle.findViewTreeLifecycleOwner
 import androidx.lifecycle.lifecycleScope
 import com.khoshnaw.ui.extenstion.runIntentInScope
 import com.khoshnaw.ui.mvi.MVIHolder
+import com.khoshnaw.viewmodel.dto.StandardStateListItem
 import com.khoshnaw.viewmodel.mvi.MVIIntent
 import com.khoshnaw.viewmodel.mvi.MVIViewModel
