--- ui/src/main/java/com/khoshnaw/ui/standard/adapter/StandardStateListItem.kt
@@ -1,7 +1,0 @@
-package com.khoshnaw.ui.standard.adapter
-
-import com.khoshnaw.ui.base.adapter.StateListItem
-
-interface StandardStateListItem : StateListItem {
-    val id: String
-}
