+++ viewModel/src/main/java/com/khoshnaw/viewmodel/dto/StandardStateListItem.kt
@@ -1,0 +1,5 @@
+package com.khoshnaw.viewmodel.dto
+
+interface StandardStateListItem : StateListItem {
+    val id: String
+}
