+++ viewModel/src/main/java/com/khoshnaw/viewmodel/dto/StateListItem.kt
@@ -1,0 +1,3 @@
+package com.khoshnaw.viewmodel.dto
+
+interface StateListItem
