+++ viewModel/src/main/java/com/khoshnaw/viewmodel/movies/MoviesState.kt
@@ -1,14 +1,14 @@
 package com.khoshnaw.viewmodel.movies
 
-import com.khoshnaw.entity.Movie
+import com.khoshnaw.viewmodel.dto.MovieUIDTO
 import com.khoshnaw.viewmodel.mvi.MVIState
 
 sealed class MoviesState(
-    open val movies: List<Movie> = listOf(),
+    open val movies: List<MovieUIDTO> = listOf(),
 ) : MVIState {
 
     class MovieList(
-        override val movies: List<Movie>,
+        override val movies: List<MovieUIDTO>,
         val isLoading: Boolean = false
     ) : MoviesState(movies)
