+++ viewModel/src/main/java/com/khoshnaw/viewmodel/movies/MoviesViewModel.kt
@@ -3,6 +3,7 @@ package com.khoshnaw.viewmodel.movies
 import com.khoshnaw.controller.movie.MovieController
 import com.khoshnaw.entity.Movie
 import com.khoshnaw.usecase.movie.loadMovieList.LoadMovieListOutputPort
+import com.khoshnaw.viewmodel.mapper.toDTO
 import com.khoshnaw.viewmodel.standard.StandardViewModel
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.Flow
@@ -27,15 +28,15 @@ class MoviesViewModel @Inject constructor(
         movieController.loadMoviesList()
     }
 
-    private suspend fun handleMovieClicked(intent: MoviesIntent.OnMovieClicked) {
+    private fun handleMovieClicked(intent: MoviesIntent.OnMovieClicked) {
         state.value?.movies?.getOrNull(intent.position)?.takeIf { it.id == intent.id }?.let {
-            movieController.showMovie(it)
+            movieController.showMovie(it.id)
         }
     }
 
     override suspend fun observeMovies(flow: Flow<List<Movie>>) {
         flow.collectResult {
-            updateState(MoviesState.MovieList(it))
+            updateState(MoviesState.MovieList(it.toDTO()))
         }
     }
